Day 4: Cleaning up rough solution + Bug Fix
My previous solution only really worked by chance -- When calculating
the sleepiest minute, I was accumulating the days being slept on on an
event-by-event basis and taking the max over this sequence.

This is incorrect -- after accumulating the days being slept on,
event-by-event, we need to group by the minutes the events occurred on
each day and take the last value in each group as the number of days the
guard was asleep on that minute in.